Backprocess fixes January 2024
A user identified a serious issue in which the way countries were assigned to daily losses (just for data files) was not working appropriately. Because countries were indexed by T64 losses, the introduction of T55/54 losses threw this off leading to unexpected behavior.

This is fixed moving forward (though at a small efficiency cost for computer processing time). I've included a new script, daily_patch, to show how this was fixed. Almost all data for 2024 has been revised accordingly.

The following series of updates will correct this issue and re-fill in back data.